requirements.txt
azure-appconfiguration==1.5.0
azure-appconfiguration-provider==1.0.0
azure-common==1.1.28
azure-core==1.29.7
azure-keyvault-secrets==4.7.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
msrest==0.7.1
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
